+++ src/pt2_rcfilters.h
@@ -5,22 +5,22 @@
 
 typedef struct onePoleFilter_t
 {
-	double tmpL, tmpR, a0, b1;
+	float tmpL, tmpR, a0, b1;
 } onePoleFilter_t;
 
 typedef struct twoPoleFilter_t
 {
-	double tmpL[4], tmpR[4], a1, a2, b1, b2;
+	float tmpL[4], tmpR[4], a1, a2, b1, b2;
 } twoPoleFilter_t;
 
 void setupOnePoleFilter(double audioRate, double cutOff, onePoleFilter_t *f);
 void clearOnePoleFilterState(onePoleFilter_t *f);
-void onePoleLPFilterStereo(onePoleFilter_t *f, const double *in, double *out);
-void onePoleHPFilterStereo(onePoleFilter_t *f, const double *in, double *out);
-void onePoleLPFilter(onePoleFilter_t *f, const double in, double *out);
-void onePoleHPFilter(onePoleFilter_t *f, const double in, double *out);
+void onePoleLPFilterStereo(onePoleFilter_t *f, const float *in, float *out);
+void onePoleHPFilterStereo(onePoleFilter_t *f, const float *in, float *out);
+void onePoleLPFilter(onePoleFilter_t *f, const float in, float *out);
+void onePoleHPFilter(onePoleFilter_t *f, const float in, float *out);
 
 void setupTwoPoleFilter(double audioRate, double cutOff, double qFactor, twoPoleFilter_t *f);
 void clearTwoPoleFilterState(twoPoleFilter_t *f);
-void twoPoleLPFilter(twoPoleFilter_t *f, const double in, double *out);
-void twoPoleLPFilterStereo(twoPoleFilter_t *f, const double *in, double *out);
+void twoPoleLPFilter(twoPoleFilter_t *f, const float in, float *out);
+void twoPoleLPFilterStereo(twoPoleFilter_t *f, const float *in, float *out);
